update in use coupons cannot be deleted

+++ web_service/src/main/java/com/codegym/web_service/Controller/employeeController/SaleManagementController.java
@@ -10,8 +10,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
+import java.util.Date;
+
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
@@ -22,11 +25,11 @@ public class SaleManagementController {
     //---------------------- list bill sprint 2 ---------------------------------
     @RequestMapping(value = "/coupon", method = RequestMethod.GET)
     public ResponseEntity<Page<Coupon>> listAllCoupon(@RequestParam("page") int page,
-                                                       @RequestParam("size") int size,
-                                                       @RequestParam("createdatefrom") String createDateFrom,
-                                                       @RequestParam("createdateto") String createDateTo,
-                                                       @RequestParam("employee") String employee,
-                                                       @RequestParam("user") String user
+                                                      @RequestParam("size") int size,
+                                                      @RequestParam("createdatefrom") String createDateFrom,
+                                                      @RequestParam("createdateto") String createDateTo,
+                                                      @RequestParam("employee") String employee,
+                                                      @RequestParam("user") String user
     ) throws ParseException {
         if (createDateFrom.equals("")) {
             createDateFrom = "1900-01-01";
@@ -47,32 +50,48 @@ public class SaleManagementController {
     @GetMapping("/coupon/{id}")
     public ResponseEntity<Coupon> getCoupon(@PathVariable("id") Integer id) {
         Coupon coupon = couponService.findById(id);
-
         return coupon == null ? new ResponseEntity<>(HttpStatus.NOT_FOUND) : new ResponseEntity<>(coupon, HttpStatus.OK);
 
     }
+
     //---------------------- Delete a Coupon ---------------------------------
     @PatchMapping("/coupon/delete/{id}")
     public ResponseEntity<Coupon> deleteCoupon(@PathVariable Integer id) {
         Coupon currentCoupon = couponService.findById(id);
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+        String date = sdf.format(new Date());
+        String couponDate = sdf.format(currentCoupon.getCreateDate());
         if (currentCoupon == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        couponService.remove(currentCoupon);
-        couponService.save(currentCoupon);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (couponDate.equals(date)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } else {
+            couponService.remove(currentCoupon);
+            couponService.save(currentCoupon);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
     }
+
 
     //---------------------- Delete Many Coupon ---------------------------------
     @GetMapping("/coupon/deletes/{id}")
     public ResponseEntity<Coupon> deleteManyCoupon(@PathVariable Integer id) {
         Coupon currentCoupon = couponService.findById(id);
+        SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd");
+        String date = sdf.format(new Date());
+        String couponDate = sdf.format(currentCoupon.getCreateDate());
         if (currentCoupon == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
-        couponService.remove(currentCoupon);
-        couponService.save(currentCoupon);
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (couponDate.equals(date)) {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        } else {
+            couponService.remove(currentCoupon);
+            couponService.save(currentCoupon);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
     }
-    //---------------------- Hieu Nguyen API - END ---------------------------------
 }
+//---------------------- Hieu Nguyen API - END ---------------------------------
+
